Move code to `x/deterministicgas`
